develop -> master

See merge request interact/ems/ems!3637

+++ jems-server/src/main/kotlin/io/cloudflight/jems/server/project/repository/report/project/resultPrinciple/ProjectReportResultPrinciplePersistenceProvider.kt
@@ -54,13 +54,13 @@ class ProjectReportResultPrinciplePersistenceProvider(
         projectResultRepository.getCumulativeValues(reportIds).toMap()
 
     @Transactional
-    override fun deleteProjectResultPrinciples(reportId: Long) {
+    override fun deleteProjectResultPrinciplesIfExist(reportId: Long) {
         val projectResults = projectResultRepository.findByProjectReportId(reportId)
-        projectResults.onEach {
-            it.attachment.deleteIfPresent()
-        }
+        projectResults.forEach { it.attachment.deleteIfPresent() }
         projectResultRepository.deleteAll(projectResults)
-        horizontalPrincipleRepository.deleteById(reportId)
+
+        if (horizontalPrincipleRepository.existsById(reportId))
+            horizontalPrincipleRepository.deleteById(reportId)
     }
 
     private fun fetchAvailablePeriodsFor(results: List<ProjectReportProjectResultEntity>) =
